+++ CHANGELOG.md
@@ -20,7 +20,8 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Removed
-- Drop the legacy registration form (#2196)
+- Remove the `logOutOneTimeAccountsAfterRegistration` setting (#2199)
+- Drop the legacy registration form (#2196, #2198, #2199)
 - Drop anything bank-data related (#2192)
 - Remove the board-related prices (#2190)
 - Drop the upgrade wizards (#2187)